requirements.txt
numpy>=1.16.0
opencv-python>= 3.4.2
pytorch>=1.8.0
scikit-image >= 0.19.2
scipy >= 1.8.1
